Move PIMPL boilerplate to the inspector base types
